refactor prebid requestBids
